Crate wabbit

This is my implementation of Wabbit, written as part of David Beazley’s excellent class on compilers.

This crate provides a Wabbit interpreter, a typechecker, a compiler that uses LLVM IR, and an export of the interpreter to WebAssembly.

The below re-exports are roughly in the order that source code is processed (branching depending on if you choose to interpret or compile).

See wabbit_rs for command-line usage

Re-exports

pub use crate::scanner::Scanner;

pub use crate::parser::Parser;

pub use crate::analyzer::Analyzer;

pub use crate::typechecker::Typechecker;

pub use crate::interpreter::Interpreter;

pub use crate::llvm::CodegenLLVM;

pub use crate::error::*;

Modules

analyzer

typecheck or interpret Wabbit AST

ast 🔒

types for Wabbit AST

environment 🔒

manage variable scope

error

error reporting

formatter

a code minimizer

interpreter

interpret Wabbit AST

llvm

generate LLVM IR

operators 🔒

Wabbit primitive operators

parser

parse Wabbit tokens

scanner

scan Wabbit source code

tokens 🔒

Wabbit tokens

typechecker

interpret Wabbit AST

types 🔒

Wabbit types and values

wasm_interpreter 🔒

a WebAssembly interpreter

Enums

Type

Wabbit data types

WabbitType

a Wabbit value, including Wabbit types

Functions

wasm_entry

entry point for WebAssembly interpreter

wasm_interp

a Wabbit interpreter, exported to WebAssembly