Enum wabbit::error::Msg

pub enum Msg {

    InvalidNumber,
    InvalidChar,
    DoubleToken,
    UnexpectedChar,
    VarDefEmpty,
    ExpectExpr,
    ExpectTypeName,
    ExpectVarName,
    ParserExpect,
    AltBranch,
    DupArgs,
    FuncDefScope,
    ReturnScope,
    AssignRetype,
    VarUndefined,
    AssignUndefined,
    ConstScope,
    InitType,
    LoopReq,
    TypeConvert,
    ConvertAirty,
    FuncAirty,
    ParamType,
    ReturnType,
    NoReturn,
    FuncUndefined,
    TypeEval,
    AccessUninit,
    ExpectType,
    TypeMatch,
    ReturnDiverge,
    RedeclareVar,
    RedeclareFunc,
    RedeclareConst,
    InternalErr,

}

collection of various errors that could be raised by an ill-formed Wabbit program

the one exception to this is Msg::InternalErr, which reports a failure of the crate itself

see Msg::msg for the messages presented to the user

Variants

InvalidNumber

InvalidChar

DoubleToken

UnexpectedChar

VarDefEmpty

ExpectExpr

ExpectTypeName

ExpectVarName

ParserExpect

AltBranch

DupArgs

FuncDefScope

ReturnScope

AssignRetype

VarUndefined

AssignUndefined

ConstScope

InitType

LoopReq

TypeConvert

ConvertAirty

FuncAirty

ParamType

ReturnType

NoReturn

FuncUndefined

TypeEval

AccessUninit

ExpectType

TypeMatch

ReturnDiverge

RedeclareVar

RedeclareFunc

RedeclareConst

InternalErr

Implementations

impl Msg

pub fn msg(&self) -> &'static str

an error message template