Trait wabbit::error::RangeReporter

pub trait RangeReporter {
    // Required method
    fn extract_tokens<'a>(&'a self, output: &mut Vec<&'a Token>);

    // Provided method
    fn extract_range(&self) -> (usize, usize) { ... }
}

trait for reporting the left and right source indices of an error

This trait is quite versatile, allowing error reporting from tokens, expressions, parsers, typecheckers, and interpreters

Required Methods

fn extract_tokens<'a>(&'a self, output: &mut Vec<&'a Token>)

extract tokens from a general type

Provided Methods

fn extract_range(&self) -> (usize, usize)

use extract_tokens to get tokens, then find the left and right edge