Struct wabbit::parser::Parser

pub struct Parser<'a> {
    current: usize,
    tokens: &'a Vec<Token>,
    statements: Vec<Stmt>,
    id: usize,
    ranges: HashMap<usize, (usize, usize)>,
}

Struct for transforming tokens into a vector of statements (AST)

Fields
current: usize

index of self.tokens that the parser is examining

tokens: &'a Vec<Token>

tokens, borrowed from a scanner

statements: Vec<Stmt>

resulting statements after parsing

id: usize

current expression/statement index

ranges: HashMap<usize, (usize, usize)>

a map from statement/expression indices to source indices

Implementations

impl<'a> Parser<'a>

pub fn new(tokens: &'a Vec<Token>) -> Self

fn assign_id<L, R>(&mut self, left: L, right: R) -> usizewhere L: RangeReporter, R: RangeReporter,

given a pair of tokens/expressions/statements, record the furthest left and right source indices

fn assign_id_single(&mut self, token: &Token) -> usize

record the source range of a single token

pub fn borrow_statements(&self) -> &Vec<Stmt>

return a reference to self.statements

pub fn borrow_ranges(&self) -> &HashMap<usize, (usize, usize)>

return a reference to self.ranges

pub fn parse(&mut self) -> Result<()>

parse all statements

fn statement(&mut self) -> Result<Stmt>

parse a single statement

This is the entry point for the mutually recursive private functions found below.

fn peek(&self) -> Token

return the current token without advancing the parser

fn previous(&mut self) -> Token

return the previous token (or current if at the first token)

fn is_end(&mut self) -> bool

check if all statements have been parsed

fn advance(&mut self) -> Token

return the current token and advance the parser one token

fn check(&mut self, t: TokenType) -> bool

check if the current token matches

fn match_any<T>(&mut self, types: T) -> boolwhere T: IntoIterator<Item = TokenType>,

check if any token is a match, and if so advance

fn expect(&mut self, c: char) -> Result<Token>

advance past and return a specific token, or return an error

fn get_name(&mut self) -> Result<Token>

advance past and return a name, or return an error

fn get_type(&mut self) -> Result<(Token, Type)>

advance past and return a type, or return an error

fn block(&mut self) -> Result<Stmt>

fn vardef(&mut self) -> Result<Stmt>

fn stmt_name(&mut self) -> Result<Stmt>

fn constdef(&mut self) -> Result<Stmt>

fn funcdef(&mut self) -> Result<Stmt>

fn loop_control(&mut self) -> Result<Stmt>

fn while_stmt(&mut self) -> Result<Stmt>

fn if_stmt(&mut self) -> Result<Stmt>

fn print_stmt(&mut self) -> Result<Stmt>

fn return_stmt(&mut self) -> Result<Stmt>

fn expr_stmt(&mut self) -> Result<Stmt>

fn expression(&mut self) -> Result<Expr>

fn or(&mut self) -> Result<Expr>

fn and(&mut self) -> Result<Expr>

fn compare(&mut self) -> Result<Expr>

fn add_or_sub(&mut self) -> Result<Expr>

fn times_or_div(&mut self) -> Result<Expr>

fn unary(&mut self) -> Result<Expr>

fn call(&mut self) -> Result<Expr>

fn finish_call(&mut self, e: Expr, lead: &Token) -> Result<Expr>

fn primary(&mut self) -> Result<Expr>