Struct wabbit::scanner::Scanner

pub struct Scanner {
    source: Vec<char>,
    tokens: Vec<Token>,
    current: usize,
    line: usize,
    start: usize,
}

Struct for transforming the raw character input of a Wabbit program into tokens

Fields
source: Vec<char>

raw character input of a Wabbit program

tokens: Vec<Token>

resulting tokens after scanning self.source

current: usize

index of self.source that the scanner is examining

line: usize

current line number the scanner is examining

start: usize

starting index before scanner gets next token

Implementations

impl Scanner

pub fn borrow_tokens(&self) -> &Vec<Token>

get a reference to a scanner’s tokens

pub fn new(s: String) -> Scanner

initialize a new scanner

fn advance(&mut self) -> char

return the current character and advance the scanner one character

fn lexeme(&self) -> String

fn add_token(&mut self, token: TokenType)

add a non-literal token to self.tokens

fn add_literal_token(&mut self, token: TokenType, l: WabbitType)

add a literal token (string, number, or identifier) to self.tokens

fn is_end(&self) -> bool

check if all characters have been scanned

fn peek(&mut self) -> char

return the current character without advancing the scanner

fn peek_next(&mut self) -> char

return the following character without advancing the scanner

fn identifier(&mut self)

scan an identifier

fn number(&mut self, found_decimal: bool) -> Result<()>

scan a numeric literal (integer or float)

fn scan_token(&mut self) -> Results<()>

scan a single token

pub fn scan(&mut self) -> Results<()>

scan all tokens